[PATCH] USB: usb-storage: handle babble
This patch introduces some handling for babble conditions.

Basically, once a babble is detected, we return sense data saying the
command was invalid.  We also go on to transfer the CSW (for BBB transport)
so we stay in phase with the device.

This isn't guaranteed to work with every device that babbles, but it can't
hurt compared to the current behavior.  Properly operating devices are
unaffected by this patch.