[PATCH] Remove DRM ioctls for common compat ioctl code
From: Andi Kleen <ak@suse.de>

This makes the 2.5 kernel with common ioctl32 code compile with DRM enabled
again.  The DRM code in the kernel is obsolete anyways and has been long
removed.  It definitely does not belong in the common ioctl emulation
layer.

Egbert Eich is working on proper 32bit DRM emulation, but it will be likely
directly integrated in the DRI/DRM sources.