[PATCH] revert adjtimex changes
From: John Stultz, George Anzinger, Eric Piel

There was confusion over the definition of TICK_USEC.  TICK_USEC is
supposed to be based on USER_HZ, however a recent change caused TICK_USEC
to be based on HZ.  This broke the adjtimex() interface on systems where
USER_HZ != HZ.  This patch reverts the change to TICK_USEC, removes an
added mis-use of the value and fixes some incorrect comments that could
lead to this sort of confusion.

Also this patch resolves the related LTP adjtimex failures.