[PATCH] remove a bogus check in sym2 driver
The sym2 driver has a check for a very broken bios where it seems
devices appear twice. This test is broken on a machine with PCI domains
since busnumber:devfn is not a unique identifier.

When we move to the scsi hotplug interface all this code will go away
but in the short term we need the following fix.