[PATCH] another timer overflow thing
in add_timer_internal() we simply leave the timer pending forever if the
expiry is in more than 0xffffffff jiffies. This means more than 48 days on
eg. ia64 - which is not an unrealistic timeout. IIRC crond is happy to use
extremely large timeouts.

It's better to time out early (if you can call 48 days "early") than to
not time out at all.