[PATCH] Fix irq handling of IO-APIC edge IRQs on UP
send_IPI_self is needed to resend irqs with IRQ_PENDING status when
enabled.

Checked with Ingo, and it's in 2.4-ac for some time.

This should fix ide lost interrupts on UP with IO-APIC

Ide trigers it this way:
 - disable_irq
 - do stuff that triggers IRQ.
 - irq is IRQ_PENDING
 - enable_irq
 - IRQ is lost, needs to be resend.

I'll send the patch to fixup IDE disable_irq logic to Bart.