[PATCH] USB: remove now-dead mode-translation code
This patch removes all of the mode-translation logic.  It's no longer
needed, as the upper-level drivers now all issue the proper 10-byte
commands.  This is a large patch, but what it's doing is removing a great
deal of unneeded code -- about 650+ lines removed, according to diffstat.

Those of you who are having some trouble getting your DVD-RAMs detected
properly should have all your problems solved by an upcoming patch from the
SCSI folks.  It may look like that problem is related to this patch, but it
isn't -- it's a side-effect of another SCSI patch to consolidate code.

Special thanks to James Bottomley for being patient enough to work with me
on these issues on the SCSI layer.

This, BTW, makes US_FL_MODE_XLATE a NOP.  The symbol is kept for the
purpose of allowing unusual_devs.h to be consistent between 2.4 and 2.5

There are still a few places where more code could (potentially) be
removed.  I need to mull over the UFI specification to be sure, tho.