o wl3501: use the regulatory domain defines
I.e. less magic numbers, also rename freq_domain to reg_domain, as
in regulatory domain, as the atmel driver does, and that made me
realize that this defines and the function that checks if a channel
is valid in a regulatory domain should be moved to the wireless
extensions (or some other place) common code.