[PATCH] Incorrect timeout in CDROM_SEND_PACKET ioctl
The CDROM_SEND_PACKET ioctl passes a struct cdrom_generic_command from
user space, which contains a timeout field.

The timeout is measured in jiffies, but the conversion from user to
kernel jiffies is missing, which makes the timeout 10 times shorter than
it should be in 2.5 kernels on x86.  This causes CDRW formatting with
cdrwtool to fail.  The following patch fixes this problem.