[PATCH] spurious SIGCHLD from dying thread group leader
A dying initial thread (thread group leader) sends SIGCHLD when it exits,
but it ought to wait until all other threads exit as well.  The cases of
secondary threads exitting first were handled properly, but not this one.

This exit.c patch fixes that test case, and I think catches the other
potential bugs of this kind as well.  The signal.c change adds some bug
catchers, the second of which will trip on the test case in the absence
of the exit.c fix.