[PATCH] USB: usb-storage: Move static string out of initializers.h
This patch seems to have fallen through the cracks.  It takes a string
constant defined as a macro in initializers.h and instead makes it a
static string in initializers.c.  It also subtracts 1 from the sizeof()
value to avoid transmitting the terminating NUL.

Finally, the patch removes an unnecessary debugging printout -- the value
printed is an _output_ buffer, so of course it won't change.